refactor(auth): one delete-account component, one form helper
The page/form split existed only so BasicDialog's confirm button could
sit outside the form, and `confirmButtonProps={{type:"submit", form}}` —
the shape SaveButton already uses — does that without one. Merging drops
FORM_ID, setDisabled, onSelfDelete, the selfDeleted ref, the bespoke
confirm button, a duplicate auth guard and an unused useUserMe()
subscription: `deleteAccount.isPending || isSuccess` derives what the ref
hand-maintained, and unlike the ref it clears itself when a delete fails,
so a failed delete no longer suppresses the login redirect.

setFieldErrors mapped v1 and allauth field errors onto form fields, and
neither branch could fire: v1 rejects a bad body as ninja's
`{"detail": [...]}`, which names no field, and the allauth branch has no
consumer — both surviving forms are v1 and single-field. What was left is
the ten lines now inline in useValidatedForm. Closes
tk-setfielderrors-v1-branch-can-never-fire-django-n-2cde7e.

The e2e page object's nested DeleteAccountForm shared the dialog's own
root locator and narrowed nothing.

Unrelated one-liner: `uv run pytest` from a worktree (CLAUDE.md's
documented invocation) creates webserver/.venv, and prettier reads only
the ignore files it is passed — so `yarn fmt-check` reported 279 files
inside the venv.

Items 23 and 24 of the #1284 review.
