+++ packages/app/src/util/forms/index.ts
@@ -1,2 +1,1 @@
-export { setFieldErrors } from "./util";
 export { useValidatedForm } from "./useValidatedForm";
